requirements.txt
flask>=3.0.0
ultralytics>=8.3.0
insightface>=0.7.3
opencv-python>=4.10.0.84
numpy>=1.24.0,<2.0
joblib>=1.3.0
onnxruntime>=1.16.0
requests>=2.31.0
mediapipe>=0.10.0

# Violation tracking dependencies
boto3==1.34.0
PyMySQL==1.1.0
python-dotenv==1.0.0
